requirements.txt
alembic==1.18.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==24.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
packaging==26.0
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.0.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.5
